+++ releaf-core/app/controllers/releaf/home_controller.rb
@@ -8,5 +8,9 @@ module Releaf
         end
       end
     end
+
+    def page_not_found
+      error_response('page_not_found', 404)
+    end
   end
 end